+++ stablehlo/dialect/VhloOps.h
@@ -23,8 +23,31 @@ limitations under the License.
 #include "mlir/IR/DialectImplementation.h"
 #include "mlir/IR/MLIRContext.h"
 #include "stablehlo/dialect/Version.h"
-#include "stablehlo/dialect/VhloDialect.h"
 #include "stablehlo/dialect/VhloTypes.h"
+
+namespace mlir {
+namespace vhlo {
+
+class VhloDialect : public Dialect {
+ public:
+  explicit VhloDialect(MLIRContext *context);
+  static StringRef getDialectNamespace() { return "vhlo"; }
+
+  // Parses a type registered to this dialect.
+  Type parseType(DialectAsmParser &parser) const override;
+
+  // Prints a type registered to this dialect.
+  void printType(Type type, DialectAsmPrinter &os) const override;
+
+  // Parses an attribute registered to this dialect.
+  Attribute parseAttribute(DialectAsmParser &parser, Type type) const override;
+
+  // Prints an attribute registered to this dialect.
+  void printAttribute(Attribute attr, DialectAsmPrinter &os) const override;
+};
+
+}  // namespace vhlo
+}  // namespace mlir
 
 // Attrs and Enums
 #include "stablehlo/dialect/VhloAttrInterfaces.h.inc"
